Make classes for handling course relationships
The old system for relationships was a nested list. It worked well for
simple THIS and THIS and (THIS and THIS), where THIS could be a single
course, or a list of interchangeable courses. It was however impossible
to handle data of the type 'two or more from this list of courses'.
To do this, I implemented CourseListPrimitive, that has quantity as a
property. This defaults to len(courses), but can be set manually to
something else. Another advatage this class gives, is the option of
giving parameters that result in courses, and not just the courses
themselves. Sometimes, you want to require some courses from a specific
faculty, or with a course in the range MAT2000 - MAT2999. This is
cumbersome if you have to list all the courses, and especially if
you want to save the data in tabular form. CourseListPrimitive
therefore only generates the actual list of courses on demand, and
its __str__ method returns a decently formatted string that doesn't
necessarily include all courses.

I also implemented CompoundCourseList, which handles lists of
CourseListPrimitives. It does so with either an 'and' or an 'or'
relationship. This allows a tree of requirements to be built up.

Both these classes have implies-methods, that takes in another object
of one of the two types. This checks if some set of requirements,
imply another.