Implement CourseList classes in search
I implemented the CourseListPrimitive and CompoundCourseList classes to
be used in the search function. After doing it, I realised that having
the str of the instances in the dataframe, and thus having to
reinstantiate classes and perform searches 4000 times to find course
requirements for a single course, was too slow. So the data is still
stored as nested lists in the dataframe, and then made into Compounds
once a match is found.